requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
gunicorn==25.0.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pyjwt==2.11.0
django_apscheduler==0.7.0
coverage==7.13.5